add: abstracted logic on hook
